refactor: useSectionValidation 훅을 useShakeAnimation 훅으로 수정
